Make prember accept the latest FastBoot manifest schema 5
`Cannot read property 'htmlFile' of undefined` is what you get when running Prember w/ Embroider, due to Embroider generating a v5 manifest, which Prember is not able to understand.

Unfortunately we seem to have to make Prember aware of manifest versions, even though the intention of [fastboot-schema.js](https://github.com/ember-fastboot/fastboot/blob/429eb1d1c3d9353f3085196e71cfde416f154cba/src/fastboot-schema.js) is to abstract this away. But as we mangle the manifest, we can not rely on that. But the change here is not really changing that.